Refactor and update readme
