+++ scripts/tsconfig.json
@@ -1,0 +1,10 @@
+{
+	"compilerOptions": {
+		"allowJs": true,
+		"checkJs": true,
+		"module": "commonjs",
+		"resolveJsonModule": true,
+		"strict": true,
+		"noEmit": true
+	}
+}
